1. Модульные тесты. [#159435]

+++ src/ru/job4j/converter/Converter.java
@@ -2,28 +2,28 @@ package ru.job4j.converter;
 
 public class Converter {
 
-    public static int rubleToEuro(int value) {
+    public static float rubleToEuro(float value) {
         return value / 70;
     }
 
-    public static int rubleToDollar(int value) {
+    public static float rubleToDollar(float value) {
         return value / 60;
     }
 
     public static void main(String[] args) {
 
-        int euro = Converter.rubleToEuro(140);
-        int dollar = Converter.rubleToDollar(120);
+        float euro = Converter.rubleToEuro(140);
+        float dollar = Converter.rubleToDollar(120);
         System.out.println("500 rubles are " + euro + " euro.");
         System.out.println("500 rubles are " + dollar + " dollar.");
         int in = 140;
         int expected = 2;
-        int out = Converter.rubleToEuro(in);
+        float out = Converter.rubleToEuro(in);
         boolean passed = expected == out;
         System.out.println("140 rubles are 2. Test result : " + passed);
         int in1 = 120;
         int expected1 = 2;
-        int out1 = Converter.rubleToDollar(in1);
+        float out1 = Converter.rubleToDollar(in1);
         boolean passed1 = expected1 == out1;
         System.out.println("120 rubles are 2. Test result : " + passed1);
     }
